replace: unsigned long => uint32_t unsigned long int => uint32_t unsigned long long => uint64_t

+++ src/apps/engine/src/s_deftab.cpp
@@ -121,8 +121,8 @@ uint32_t S_DEFTAB::MakeHashValue(const char *string)
         auto v = *string++;
         if ('A' <= v && v <= 'Z')
             v += 'a' - 'A'; // case independent
-        hval = (hval << 4) + static_cast<unsigned long>(v);
-        const uint32_t g = hval & (static_cast<unsigned long>(0xf) << (32 - 4));
+        hval = (hval << 4) + static_cast<uint32_t>(v);
+        const uint32_t g = hval & (static_cast<uint32_t>(0xf) << (32 - 4));
         if (g != 0)
         {
             hval ^= g >> (32 - 8);
