+++ src/libs/location/src/locator_array.cpp
@@ -153,14 +153,14 @@ long LocatorArray::FindByName(const char *locName)
 
 long LocatorArray::CalcHashString(const char *str)
 {
-    unsigned long hval = 0;
+    uint32_t hval = 0;
     while (*str != '\0')
     {
         auto c = *str++;
         if (c >= 'A' && c <= 'Z')
             c += 'a' - 'A';
-        hval = (hval << 4) + static_cast<unsigned long>(c);
-        const auto g = hval & (static_cast<unsigned long>(0xf) << (32 - 4));
+        hval = (hval << 4) + static_cast<uint32_t>(c);
+        const auto g = hval & (static_cast<uint32_t>(0xf) << (32 - 4));
         if (g != 0)
         {
             hval ^= g >> (32 - 8);
